+++ CHANGELOG.md
@@ -6,9 +6,10 @@ All notable changes to this project will be documented in this file.
 
 ### New
 
-- `config` config parameter allows to specify configuration via JSON-based config file.
-- port all code to the type script.
+- `config` config parameter allows to specify configuration via JSON-based config file
+- config file reload without restart on `SIGHUP` signal
 - optimizations for aggregation queries `MIN` and `MAX` – use simple sorted queries instead of `COLLECT AGGREGATE`.
+- port all code to the type script.
 
 ## [0.40.0] 2021-06-28